classifier/requirements.txt
coloredlogs
jsonpath-ng
jsonschema
Levenshtein
MarkupSafe
PyYAML
regex
requests
cython
pyyaml
ruamel.yaml
langchain
chromadb
sentence-transformers
tensorboardX
tensorboard
addict
